Add a bunch of missing static_cast<float>()

+++ src/mavsdk/plugins/winch/winch_impl.cpp
@@ -145,7 +145,7 @@ void WinchImpl::relative_length_control_async(
     MavlinkCommandSender::CommandLong command{};
 
     command.command = MAV_CMD_DO_WINCH;
-    command.params.maybe_param1 = instance;
+    command.params.maybe_param1 = static_cast<float>(instance);
     command.params.maybe_param2 = static_cast<float>(Winch::WinchAction::RelativeLengthControl);
     command.params.maybe_param3 = length;
     command.params.maybe_param4 = rate;
@@ -175,7 +175,7 @@ void WinchImpl::rate_control_async(
     MavlinkCommandSender::CommandLong command{};
 
     command.command = MAV_CMD_DO_WINCH;
-    command.params.maybe_param1 = instance;
+    command.params.maybe_param1 = static_cast<float>(instance);
     command.params.maybe_param2 = static_cast<float>(Winch::WinchAction::RateControl);
     command.params.maybe_param4 = rate;
 
@@ -202,7 +202,7 @@ void WinchImpl::lock_async(uint32_t instance, const Winch::ResultCallback callba
     MavlinkCommandSender::CommandLong command{};
 
     command.command = MAV_CMD_DO_WINCH;
-    command.params.maybe_param1 = instance;
+    command.params.maybe_param1 = static_cast<float>(instance);
     command.params.maybe_param2 = static_cast<float>(Winch::WinchAction::Lock);
 
     command.target_component_id = MAV_COMPONENT::MAV_COMP_ID_WINCH;
@@ -228,7 +228,7 @@ void WinchImpl::deliver_async(uint32_t instance, const Winch::ResultCallback cal
     MavlinkCommandSender::CommandLong command{};
 
     command.command = MAV_CMD_DO_WINCH;
-    command.params.maybe_param1 = instance;
+    command.params.maybe_param1 = static_cast<float>(instance);
     command.params.maybe_param2 = static_cast<float>(Winch::WinchAction::Deliver);
 
     command.target_component_id = MAV_COMPONENT::MAV_COMP_ID_WINCH;
@@ -254,7 +254,7 @@ void WinchImpl::hold_async(uint32_t instance, const Winch::ResultCallback callba
     MavlinkCommandSender::CommandLong command{};
 
     command.command = MAV_CMD_DO_WINCH;
-    command.params.maybe_param1 = instance;
+    command.params.maybe_param1 = static_cast<float>(instance);
     command.params.maybe_param2 = static_cast<float>(Winch::WinchAction::Hold);
 
     command.target_component_id = MAV_COMPONENT::MAV_COMP_ID_WINCH;
@@ -280,7 +280,7 @@ void WinchImpl::retract_async(uint32_t instance, const Winch::ResultCallback cal
     MavlinkCommandSender::CommandLong command{};
 
     command.command = MAV_CMD_DO_WINCH;
-    command.params.maybe_param1 = instance;
+    command.params.maybe_param1 = static_cast<float>(instance);
     command.params.maybe_param2 = static_cast<float>(Winch::WinchAction::Retract);
 
     command.target_component_id = MAV_COMPONENT::MAV_COMP_ID_WINCH;
@@ -306,7 +306,7 @@ void WinchImpl::load_line_async(uint32_t instance, const Winch::ResultCallback c
     MavlinkCommandSender::CommandLong command{};
 
     command.command = MAV_CMD_DO_WINCH;
-    command.params.maybe_param1 = instance;
+    command.params.maybe_param1 = static_cast<float>(instance);
     command.params.maybe_param2 = static_cast<float>(Winch::WinchAction::LoadLine);
 
     command.target_component_id = MAV_COMPONENT::MAV_COMP_ID_WINCH;
@@ -332,7 +332,7 @@ void WinchImpl::abandon_line_async(uint32_t instance, const Winch::ResultCallbac
     MavlinkCommandSender::CommandLong command{};
 
     command.command = MAV_CMD_DO_WINCH;
-    command.params.maybe_param1 = instance;
+    command.params.maybe_param1 = static_cast<float>(instance);
     command.params.maybe_param2 = static_cast<float>(Winch::WinchAction::AbandonLine);
 
     command.target_component_id = MAV_COMPONENT::MAV_COMP_ID_WINCH;
@@ -358,7 +358,7 @@ void WinchImpl::load_payload_async(uint32_t instance, const Winch::ResultCallbac
     MavlinkCommandSender::CommandLong command{};
 
     command.command = MAV_CMD_DO_WINCH;
-    command.params.maybe_param1 = instance;
+    command.params.maybe_param1 = static_cast<float>(instance);
     command.params.maybe_param2 = static_cast<float>(Winch::WinchAction::LoadPayload);
 
     command.target_component_id = MAV_COMPONENT::MAV_COMP_ID_WINCH;
